+++ goji/client.py
@@ -14,6 +14,7 @@ from goji.models import (
     Comment,
     Comments,
     Issue,
+    IssueLinkType,
     SearchResults,
     Sprint,
     Transition,
@@ -127,6 +128,24 @@ class JIRAClient(object):
         )
         self.validate_response(response)
         return list(map(Attachment.from_json, response.json()))
+
+    def get_issue_link_types(self) -> List[IssueLinkType]:
+        response = self.get('issueLinkType')
+        return list(map(IssueLinkType.from_json, response.json()['issueLinkTypes']))
+
+    def link_issue(self, outward_issue: str, inward_issue: str, type: str):
+        data = {
+            'type': {
+                'name': type,
+            },
+            'inwardIssue': {
+                'key': inward_issue,
+            },
+            'outwardIssue': {
+                'key': outward_issue,
+            },
+        }
+        self.post('issueLink', data)
 
     def create_issue(self, fields) -> Issue:
         response = self.post('issue', {'fields': fields})