+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `goji search` contains an `--all` option to view all issues (paginating
   through each page).
 - Added `goji link` command to link two issues together.
+- Add support for an undocumented and unstable plugin ecosystem to extend goji.
 
 ## 0.5.0