+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # GOJI Changelog
 
-## TBD
+## 0.6.0 (2025/03/10)
 
 ### Enhancements